pyproject.toml
[tool.poetry]
name = "wooyaa-puzzlebot-integrator"
version = "0.1.0"
description = ""
authors = ["katrin_texet <lediaevak2001@mail.ru>"]
readme = "README.md"
packages = [{include = "wooyaa_puzzlebot_integrator"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.8"
pyaml-env = "^1.2.1"
sqlalchemy = "^2.0.19"
fastapi = "^0.101.1"
alembic = "^1.11.2"
more-itertools = "^10.1.0"
psycopg2-binary = "^2.9.7"
uvicorn = "^0.23.2"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
httpx = "^0.24.1"
pyjwt = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
